Move `TokenLattice` and `CharTrie` to data-structures module
