+++ src/tokenizers.js
@@ -35,7 +35,11 @@ import {
 import { max, min, round } from './utils/maths.js';
 import { Tensor } from './utils/tensor.js';
 
-import { PriorityQueue } from './utils/data-structures.js';
+import {
+    PriorityQueue,
+    TokenLattice,
+    CharTrie,
+ } from './utils/data-structures.js';
 
 /**
  * @typedef {import('./utils/hub.js').PretrainedOptions} PretrainedOptions
@@ -3391,231 +3395,6 @@ export class MarianTokenizer extends PreTrainedTokenizer {
         }
     }
 
-}
-
-/**
- * A trie structure to efficiently store and search for strings.
- */
-class CharTrie {
-    constructor() {
-        this.root = CharTrieNode.default();
-    }
-
-    /**
-     * Adds one or more `texts` to the trie.
-     * @param {string[]} texts The strings to add to the trie.
-     */
-    extend(texts) {
-        for (let text of texts) {
-            this.push(text);
-        }
-    }
-
-    /**
-     * Adds one or more `texts` to the trie.
-     * @param {*} text The strings to add to the trie.
-     */
-    push(text) {
-        let node = this.root;
-        for (let ch of text) {
-            let child = node.children.get(ch);
-            if (child === undefined) {
-                child = CharTrieNode.default();
-                node.children.set(ch, child);
-            }
-            node = child;
-        }
-        node.isLeaf = true;
-    }
-
-    /**
-     * Searches the trie for all strings with a common prefix of `text`.
-     * @param {string} text The common prefix to search for.
-     * @yields {string} Each string in the trie that has `text` as a prefix.
-     */
-    *commonPrefixSearch(text) {
-        let node = this.root;
-        let prefix = "";
-        for (let i = 0; i < text.length && node !== undefined; ++i) {
-            const ch = text[i];
-            prefix += ch;
-            node = node.children.get(ch);
-            if (node !== undefined && node.isLeaf) {
-                yield prefix;
-            }
-        }
-    }
-}
-
-/**
- * Represents a node in a character trie.
- * @param {boolean} isLeaf Whether the node is a leaf node or not.
- * @param {Map<string, CharTrieNode>} children A map containing the node's children, where the key is a character and the value is a `CharTrieNode`.
- */
-class CharTrieNode {
-    constructor(isLeaf, children) {
-        this.isLeaf = isLeaf;
-        this.children = children;
-    }
-
-    /**
-     * Returns a new `CharTrieNode` instance with default values.
-     * @returns {CharTrieNode} A new `CharTrieNode` instance with `isLeaf` set to `false` and an empty `children` map.
-     */
-    static default() {
-        return new CharTrieNode(false, new Map());
-    }
-}
-
-class TokenLattice {
-    /**
-     * Creates a new TokenLattice instance.
-     *
-     * @param {string} sentence The input sentence to be tokenized.
-     * @param {number} bosTokenId The beginning-of-sequence token ID.
-     * @param {number} eosTokenId The end-of-sequence token ID.
-     */
-    constructor(sentence, bosTokenId, eosTokenId) {
-        this.sentence = sentence;
-        this.len = sentence.length;
-        this.bosTokenId = bosTokenId;
-        this.eosTokenId = eosTokenId;
-        this.nodes = [];
-        this.beginNodes = new Array(this.len + 1);
-        this.endNodes = new Array(this.len + 1);
-        for (let i = 0; i < this.len + 1; ++i) {
-            this.beginNodes[i] = [];
-            this.endNodes[i] = [];
-        }
-        const bos = new TokenLatticeNode(this.bosTokenId, 0, 0, 0, 0.0);
-        const eos = new TokenLatticeNode(this.eosTokenId, 1, this.len, 0, 0.0);
-        this.nodes.push(bos.clone());
-        this.nodes.push(eos.clone());
-        this.beginNodes[this.len].push(eos);
-        this.endNodes[0].push(bos);
-    }
-
-    /**
-     * Inserts a new token node into the token lattice.
-     *
-     * @param {number} pos The starting position of the token.
-     * @param {number} length The length of the token.
-     * @param {number} score The score of the token.
-     * @param {number} tokenId The token ID of the token.
-     */
-    insert(pos, length, score, tokenId) {
-        const nodeId = this.nodes.length;
-        const node = new TokenLatticeNode(tokenId, nodeId, pos, length, score);
-        this.beginNodes[pos].push(node);
-        this.endNodes[pos + length].push(node);
-        this.nodes.push(node);
-    }
-
-    /**
-     * Implements the Viterbi algorithm to compute the most likely sequence of tokens.
-     *
-     * @returns {TokenLatticeNode[]} The array of nodes representing the most likely sequence of tokens.
-     */
-    viterbi() {
-        const len = this.len;
-        let pos = 0;
-        while (pos <= len) {
-            if (this.beginNodes[pos].length == 0) {
-                return [];
-            }
-            for (let rnode of this.beginNodes[pos]) {
-                rnode.prev = null;
-                let bestScore = 0.0;
-                let bestNode = null;
-                for (let lnode of this.endNodes[pos]) {
-                    const score = lnode.backtraceScore + rnode.score;
-                    if (bestNode === null || score > bestScore) {
-                        bestNode = lnode.clone();
-                        bestScore = score;
-                    }
-                }
-
-                if (bestNode !== null) {
-                    rnode.prev = bestNode;
-                    rnode.backtraceScore = bestScore;
-                } else {
-                    return [];
-                }
-            }
-            ++pos;
-        }
-
-        const results = [];
-        const root = this.beginNodes[len][0];
-        const prev = root.prev;
-        if (prev === null) {
-            return [];
-        }
-
-        let node = prev.clone();
-        while (node.prev !== null) {
-            results.push(node.clone());
-            const n = node.clone();
-            node = n.prev.clone();
-        }
-
-        results.reverse();
-        return results;
-    }
-
-    /**
-     * @param {TokenLatticeNode} node
-     * @returns {string} The array of nodes representing the most likely sequence of tokens.
-     */
-    piece(node) {
-        return this.sentence.slice(node.pos, node.pos + node.length);
-    }
-
-    /**
-     * @returns {Array} The array of nodes representing the most likely sequence of tokens.
-     */
-    tokens() {
-        const nodes = this.viterbi();
-        return nodes.map(x => this.piece(x));
-    }
-
-    /**
-     * @returns {Array} The array of nodes representing the most likely sequence of tokens.
-     */
-    tokenIds() {
-        const nodes = this.viterbi();
-        return nodes.map(x => x.tokenId);
-    }
-}
-class TokenLatticeNode {
-    /**
-     * Represents a node in a token lattice for a given sentence.
-     * @param {number} tokenId The ID of the token associated with this node.
-     * @param {number} nodeId The ID of this node.
-     * @param {number} pos The starting position of the token in the sentence.
-     * @param {number} length The length of the token.
-     * @param {number} score The score associated with the token.
-     */
-    constructor(tokenId, nodeId, pos, length, score) {
-        this.tokenId = tokenId;
-        this.nodeId = nodeId;
-        this.pos = pos;
-        this.length = length;
-        this.score = score;
-        this.prev = null;
-        this.backtraceScore = 0.0;
-    }
-
-    /**
-     * Returns a clone of this node.
-     * @returns {TokenLatticeNode} A clone of this node.
-     */
-    clone() {
-        const n = new TokenLatticeNode(this.tokenId, this.nodeId, this.pos, this.length, this.score);
-        n.prev = this.prev;
-        n.backtraceScore = this.backtraceScore;
-        return n;
-    }
 }
 
 /**
